#27 - moves utils to ui module
