+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/recurringexpenses/RecurringExpensesScreen.kt
@@ -66,7 +66,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import br.com.rbrthmn.R
-import br.com.rbrthmn.ui.financialcompanion.utils.valueWithCurrencyString
+import br.com.rbrthmn.ui.utils.valueWithCurrencyString
 
 data class RecurringExpense(val description: String, val value: String, val billingDay: String)
