+++ core/ui/src/main/java/br/com/rbrthmn/ui/utils/StringProvider.kt
@@ -18,10 +18,17 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.utils
+package br.com.rbrthmn.ui.utils
 
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.res.stringResource
+import android.content.Context
+import androidx.annotation.StringRes
 
-@Composable
-fun valueWithCurrencyString(currencyStringId: Int, value: String) = "${stringResource(id = currencyStringId)} $value"
+interface StringProvider {
+    fun getString(@StringRes stringId: Int): String
+}
+
+class ResourceStringProvider(private val context: Context) : StringProvider {
+    override fun getString(@StringRes stringId: Int): String {
+        return context.getString(stringId)
+    }
+}
